refactor: split utils.ts into domain modules
Extract 6 domain-specific modules from monolithic utils.ts:
- cn.ts: Tailwind class merge
- format.ts: capitalize, formatTokens
- math.ts: minMax
- audio.ts: calculateAudioDuration
- file.ts: createUtf8SubtitleBlob
- async.ts: sleep

utils.ts is now a re-export barrel — 87 import sites unchanged.
