+++ src/lib/utils.ts
@@ -1,64 +1,6 @@
-import { SubtitleType } from "@/types/subtitles"
-import { clsx, type ClassValue } from "clsx"
-import { twMerge } from "tailwind-merge"
-
-export function cn(...inputs: ClassValue[]) {
-  return twMerge(clsx(inputs))
-}
-
-export function minMax(value: number, min: number, max: number): number {
-  return Math.min(Math.max(value, min), max)
-}
-
-export function sleep(ms: number): Promise<void> {
-  return new Promise((resolve) => setTimeout(resolve, ms))
-}
-
-export function capitalize(s: string): string {
-  if (!s) return s
-  const arr = s.split(" ")
-  arr.forEach((val, i) => arr[i] = val[0].toUpperCase() + val.slice(1))
-  return arr.join(" ")
-}
-
-export function formatTokens(tokens: number | undefined): string {
-  if (tokens === undefined) {
-    return ""
-  }
-  if (tokens >= 1_000_000) {
-    const value = (tokens / 1_000_000).toFixed(1)
-    const intVal = parseInt(value)
-    return `${Number(value) - intVal ? value : intVal}M tokens`
-  }
-  if (tokens >= 1_000) {
-    const value = tokens / 1_000
-    return `${value.toFixed(0)}k tokens`
-  }
-  return `${tokens} tokens`
-}
-
-export function calculateAudioDuration(file: File): Promise<number> {
-  return new Promise((resolve) => {
-    const audio = new Audio()
-    const objectUrl = URL.createObjectURL(file)
-
-    audio.addEventListener('loadedmetadata', () => {
-      URL.revokeObjectURL(objectUrl)
-      resolve(audio.duration || 0)
-    })
-
-    audio.addEventListener('error', () => {
-      URL.revokeObjectURL(objectUrl)
-      resolve(0)
-    })
-
-    audio.src = objectUrl
-  })
-}
-
-export function createUtf8SubtitleBlob(content: string, type: SubtitleType): Blob {
-  if (type === "vtt") {
-    return new Blob([content], { type: "text/vtt;charset=utf-8" })
-  }
-  return new Blob(["\ufeff", content], { type: "text/plain;charset=utf-8" })
-}
+export { cn } from "./utils/cn"
+export { minMax } from "./utils/math"
+export { sleep } from "./utils/async"
+export { capitalize, formatTokens } from "./utils/format"
+export { calculateAudioDuration } from "./utils/audio"
+export { createUtf8SubtitleBlob } from "./utils/file"
