+++ src/lib/utils/math.ts
@@ -1,0 +1,3 @@
+export function minMax(value: number, min: number, max: number): number {
+  return Math.min(Math.max(value, min), max)
+}
